# docs/API-SPEC.yaml
openapi: 3.1.0
info:
  title: DivalHR API
  version: 0.1.0
  description: Initial contract skeleton for the bilingual DivalHR platform.
...
servers:
  - url: https://api.divalhr.com/api/v1
security:
  - oauth2: []
paths:
...
  /employees/{employeeId}:
    get:
      operationId: getEmployee
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: tenant-admin
      x-divalhr-required-assurance: mfa
      summary: Read one employee's profile and current employment state
      description: >-
        MVP-021. The employee, the employment period and, per assignment kind, the row covering
        the business date (today in the organization's time zone), or null. Unknown, foreign and
        malformed IDs give the same 404 EMPLOYEE_NOT_FOUND. Recorded as a disclosure (audit
        employee.read). Cache-Control private, no-store.
      tags: [people]
      parameters:
        - $ref: "#/components/parameters/EmployeeId"
        - $ref: "#/components/parameters/CorrelationId"
      responses:
        "200":
          description: The employee profile
          headers:
            Cache-Control:
              $ref: "#/components/headers/CacheControlPrivateNoStore"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/PrivilegedForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /employees/{employeeId}/contracts:
    get:
      operationId: listEmployeeContracts
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: tenant-admin
      x-divalhr-required-assurance: mfa
      summary: List an employee's contracts
      description: MVP-030. Contracts newest first, without content. Recorded as a disclosure (audit contract.read).
        Cache-Control private, no-store.
      tags:
      - documents
      parameters:
      - $ref: '#/components/parameters/EmployeeId'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/ContractLimit'
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: OK
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlPrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/PrivilegedForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    post:
      operationId: issueEmployeeContract
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: tenant-admin
      x-divalhr-required-assurance: mfa
      summary: Issue a contract
      description: MVP-030. Repeats a preview with the expectedEmploymentVersion and previewDigest it returned.
        Under the employment lock everything is re-rendered and the digest recomputed (409 CONTRACT_PREVIEW_CHANGED
        on any difference); the immutable snapshot, audit and outbox commit together. Employment history
        is never changed.
      tags:
      - documents
      parameters:
      - $ref: '#/components/parameters/EmployeeId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueContract'
      responses:
        '201':
          description: Issued, or the original issue replayed
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/PrivilegedForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /employees/{employeeId}/contracts/preview:
    post:
      operationId: previewEmployeeContract
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: tenant-admin
      x-divalhr-required-assurance: mfa
      summary: Preview a contract
      description: MVP-030. Renders an approved template version for the employee's current employment and
        returns the exact snapshot, warnings and the preview digest. Writes nothing but the disclosure audit
        (contract.preview).
      tags:
      - documents
      parameters:
      - $ref: '#/components/parameters/EmployeeId'
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractPreviewCommand'
      responses:
        '200':
          description: OK
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlPrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractPreview'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/PrivilegedForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /employees/{employeeId}/contracts/{contractId}:
    get:
      operationId: getEmployeeContract
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: tenant-admin
      x-divalhr-required-assurance: mfa
      summary: Read an employee's contract
      description: MVP-030. The contract with its snapshot and acknowledgement evidence. Recorded as a disclosure
        (audit contract.read).
      tags:
      - documents
      parameters:
      - $ref: '#/components/parameters/EmployeeId'
      - $ref: '#/components/parameters/ContractId'
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: OK
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlPrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/PrivilegedForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /employees/{employeeId}/contracts/{contractId}/void:
    post:
      operationId: voidEmployeeContract
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: tenant-admin
      x-divalhr-required-assurance: mfa
      summary: Void an unacknowledged contract
      description: MVP-030. Voids an ISSUED contract (expectedVersion, closed reason). An acknowledged contract
        is never voided. Content and digests are kept.
      tags:
      - documents
      parameters:
      - $ref: '#/components/parameters/EmployeeId'
      - $ref: '#/components/parameters/ContractId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoidContract'
      responses:
        '200':
          description: OK
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
            Cache-Control:
              $ref: '#/components/headers/CacheControlPrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contract'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/PrivilegedForbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /me/contracts/{contractId}:
    get:
      operationId: getMyContract
      x-divalhr-lifecycle: implemented
      x-divalhr-scope: tenant
      x-divalhr-required-role: employee
      summary: Read my contract
      description: MVP-030. One of the caller's own contracts with its snapshot, the acknowledgement statements
        in French and English with their digests, and the evidence once acknowledged. Another employee's
        or tenant's contract is the same 404 as an unknown one.
      tags:
      - documents
      parameters:
      - $ref: '#/components/parameters/ContractId'
      - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: OK
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlPrivateNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyContract'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://identity.divalhr.com/oauth2/authorize
          tokenUrl: https://identity.divalhr.com/oauth2/token
          scopes:
            organizations:read: Read authorized organizations
            employees:read: Read authorized employees
            employees:write: Manage authorized employees
            attendance:write: Record attendance
            payroll:read: Read authorized payroll data
  headers:
...
    IdempotentReplayed:
      description: Present with value "true" when the response replays an earlier successful request.
      schema:
        type: string
        enum: ["true"]
...
    CacheControlPrivateNoStore:
      description: >-
        Invitation administration responses may contain email addresses; browsers, offline caches
        and intermediaries must not retain them.
      schema:
        type: string
        enum: ["private, no-store"]
...
  responses:
    Unauthorized:
      description: Missing, invalid or expired access token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    PrivilegedForbidden:
      description: >-
        Authenticated but not permitted. Privileged operations (x-divalhr-required-assurance: mfa)
        additionally return MFA_REQUIRED when the verified access token's acr claim is not exactly
        urn:divalhr:loa:mfa (multifactor authentication not completed for this session). Checks run
        in this order before the request is read: verified subject, required role (ACCESS_DENIED),
        verified tenant (TENANT_CONTEXT_MISSING), then MFA assurance (MFA_REQUIRED), then, for
        tenant-scoped operations, a matching tenant membership (MVP-012A): the caller's membership
        must belong to the token's tenant and carry exactly the required role, which the token must
        also hold. A missing or non-matching membership returns the same ACCESS_DENIED as a missing
        role, with no further detail. MFA_REQUIRED carries no params and never echoes claims,
        roles, subject, tenant or request data.
      headers:
        WWW-Authenticate:
          $ref: "#/components/headers/WwwAuthenticateStepUp"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    BadRequest:
      description: >-
        Invalid request. VALIDATION_FAILED lists params.fields[{field, constraint}] with
        constraint REQUIRED, LENGTH, FORMAT, RANGE, DUPLICATE or UNKNOWN_PROPERTY. The
        *_NOT_SUPPORTED codes carry params.field and params.supported. EFFECTIVE_DATE_INVALID,
        SITE_PERIOD_OUTSIDE_LEGAL_ENTITY, DEPARTMENT_PERIOD_OUTSIDE_SITE,
        COST_CENTER_PERIOD_OUTSIDE_SITE, REGION_PERIOD_OUTSIDE_LEGAL_ENTITY,
        TEAM_PERIOD_OUTSIDE_DEPARTMENT and TEAM_PERIOD_OUTSIDE_COST_CENTER carry params.field
        (effectiveFrom or effectiveTo). TEAM_PARENT_REQUIRED (no parent supplied) and
        TEAM_PARENT_AMBIGUOUS (both parents supplied) carry no params and follow format
        validation. SITE_PERIOD_OUTSIDE_REGION and
        SITE_REGION_LEGAL_ENTITY_MISMATCH carry params.field = regionId. CURSOR_INVALID carries no
        params. Invitation fields (email, role, locale) use VALIDATION_FAILED; an unsupported role,
        including platform-admin, is constraint FORMAT exactly like any unknown value. Submitted
        values, and email addresses in particular, are never echoed. IMPORT_FILE_INVALID (MVP-020)
        carries params.reason (EMPTY, ENCODING, DELIMITER, MALFORMED, LINE_TOO_LONG,
        TOO_MANY_COLUMNS, COLUMN_UNKNOWN, COLUMN_DUPLICATE, COLUMN_MISSING, NO_ROWS or
        TOO_MANY_ROWS) and, for header problems, params.column (1-based header position, or the
        missing column key for COLUMN_MISSING); the file content is never echoed. A request body
        that is not text/csv (multipart included) is VALIDATION_FAILED on field Content-Type.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: >-
        IDEMPOTENCY_KEY_REUSED (the key was already used with a different payload), or
        DUPLICATE_LEGAL_ENTITY_CODE / DUPLICATE_SITE_CODE / DUPLICATE_DEPARTMENT_CODE /
        DUPLICATE_COST_CENTER_CODE / DUPLICATE_REGION_CODE / DUPLICATE_TEAM_CODE (the code
        already exists for that
        resource type in the tenant, in any letter case), or SITE_REGION_ALREADY_ASSIGNED (the site
        already has a different region; carries no params), or INVITATION_ALREADY_PENDING /
        INVITATION_RECIPIENT_ALREADY_MEMBER (params.field = email; only the caller's own tenant is
        consulted), INVITATION_NOT_PENDING (the invitation is accepted, expired or being
        accepted), INVITATION_CANNOT_BE_ACCEPTED or INVITATION_ACCEPTANCE_IN_PROGRESS (no params),
        or TENANT_ADMIN_BOOTSTRAP_UNAVAILABLE (the organization already has a tenant administrator
        or an open tenant-admin invitation; no params, and which of the two is not disclosed), or,
        for employee imports (MVP-020), IMPORT_LIMIT_REACHED (3 open imports in the tenant),
        IMPORT_NOT_COMMITTABLE (params.status: the import is no longer open), IMPORT_PREVIEW_CHANGED
        (the commit does not repeat the staged preview), IMPORT_STALE (a row became invalid since
        the preview; nothing was created) or IMPORT_NOTHING_TO_COMMIT (no valid row), or, for
        employment history (MVP-021), EMPLOYMENT_VERSION_CONFLICT (the employment changed since the
        preview), EMPLOYMENT_PREVIEW_CHANGED (the recomputed change differs from the preview),
        EMPLOYMENT_CHANGE_DATE_TAKEN (the kind already changes on that date),
        EMPLOYMENT_CHANGE_NOT_CANCELLABLE (not an active future change) or
        EMPLOYMENT_CHANGE_HAS_DEPENDENTS (a later row derives from the change), or, for
        separations (MVP-022), SEPARATION_EXISTS (the employment already has a separation that is
        not cancelled), SEPARATION_FUTURE_CHANGES (future effects after the last day must be
        cancelled first; params.count), SEPARATION_PROTECTED (params.reason SELF or ADMIN_ACCESS),
        SEPARATION_PREVIEW_CHANGED, SEPARATION_NOT_CANCELLABLE (params.reason NOT_SCHEDULED,
        ACCESS_ALREADY_REVOKED or HISTORY_CHANGED_SINCE), SEPARATION_TASK_VERSION_CONFLICT,
        SEPARATION_TASK_CLOSED, ACCESS_LINK_CONFLICT (params.reason EMPLOYEE_LINKED,
        ALREADY_LINKED or ACCESS_REVOKED), ACCESS_LINK_LOCKED, ACCESS_LINK_VERSION_CONFLICT or
        ACCESS_REVOCATION_NOT_RETRYABLE. Other MVP-021/022 codes carry no params.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: >-
        LEGAL_ENTITY_NOT_FOUND, SITE_NOT_FOUND, REGION_NOT_FOUND, DEPARTMENT_NOT_FOUND,
        COST_CENTER_NOT_FOUND, INVITATION_NOT_FOUND, EMPLOYEE_IMPORT_NOT_FOUND, EMPLOYEE_NOT_FOUND,
        EMPLOYMENT_CHANGE_NOT_FOUND, SEPARATION_NOT_FOUND, SEPARATION_TASK_NOT_FOUND or
        MEMBERSHIP_NOT_FOUND (MVP-022; also for an address with no membership) - the referenced
        resource does not exist in the caller's tenant. Missing and foreign-tenant resources are
        indistinguishable. ORGANIZATION_NOT_FOUND (platform operations) - the organization ID is
        malformed, unknown or not active; the cases are indistinguishable and the ID is never
        echoed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TooManyRequests:
      description: >-
        INVITATION_RATE_LIMITED (the tenant's invitation quota is used up),
        INVITATION_RESEND_LIMITED (at most 3 reissues per invitation, at least 5 minutes apart) or
        RATE_LIMITED (too many anonymous requests from one client, too many bootstrap
        invitations created by one platform administrator, more than 30 access-review requests
        per minute by one subject, or, for employee imports, more than 10 upload and commit
        requests per minute by one subject, 30 uploads or 60 commits per 10 minutes in one tenant,
        or, for employees and employment history, more than 60 reads, searches and previews or 20
        changes and cancellations per minute by one subject, 300 searches or 200 changes and
        cancellations per 10 minutes in one tenant).
        Limits are per instance. No params.
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ServiceUnavailable:
      description: >-
        IDENTITY_PROVIDER_UNAVAILABLE - the identity provider could not be reached; retry with the
        same token. IMPORT_TIMEOUT - an employee import's database work exceeded its time limit and
        was rolled back entirely; retry with the same Idempotency-Key. EMPLOYMENT_CHANGE_TIMEOUT -
        an employment change or cancellation, a separation, its cancellation or an access-link
        change exceeded its time limit or met a lock held by a background job, and nothing was
        written; retry with the same Idempotency-Key. No params.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    UnprocessableContent:
      description: >-
        MVP-021 business rules. EMPLOYMENT_CHANGE_NO_EFFECT (params.field: the assignment kind whose
        value would not change), EMPLOYMENT_DATE_OUTSIDE_EMPLOYMENT, RETROACTIVE_WINDOW_EXCEEDED
        (before the configurable pilot window, 60 days by default), MANAGER_INVALID (params.reason:
        NOT_FOUND, SELF, NOT_EMPLOYED, CYCLE or CHAIN_TOO_DEEP) or PLACEMENT_INVALID (params.field:
        legalEntityId, siteId, departmentId, costCenterId or teamId; params.reason: NOT_FOUND,
        MISMATCH, NOT_EFFECTIVE or ENDS_DURING_PERIOD). MVP-022: SEPARATION_DATE_OUT_OF_RANGE
        (before the employment start, the retroactive window, or more than 180 days ahead),
        SEPARATION_ACCESS_TIMING_INVALID (IMMEDIATELY with a future last day),
        SEPARATION_ACKNOWLEDGEMENT_REQUIRED (params.acknowledgement),
        SEPARATION_REPORT_PLAN_REQUIRED, SEPARATION_TOO_MANY_INTERVALS (params.count) and
        MANAGER_INVALID for the replacement manager. Params never carry a date, a value or an ID.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  parameters:
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Optional caller-supplied ID (8-64 chars of A-Z a-z 0-9 . _ -); otherwise generated.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{8,64}$"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Client-generated key, scoped to the operation and the authenticated subject. Records are
        retained for at least 7 days and honoured until cleanup removes them; cleanup never removes
        a record before its retention boundary.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: "^[A-Za-z0-9._:-]{16,128}$"
...
    Cursor:
      name: cursor
      in: query
      description: >-
        Opaque, signed continuation token from a previous page's nextCursor. It is bound to the
        operation, tenant and filters that produced it; any other use returns CURSOR_INVALID.
      schema:
        type: string
        maxLength: 512
...
    EmployeeId:
      name: employeeId
      in: path
      required: true
      description: Malformed, unknown and foreign IDs give the same 404 EMPLOYEE_NOT_FOUND.
      schema:
        type: string
...
    ContractId:
      name: contractId
      in: path
      required: true
      description: Malformed, unknown, foreign and other-employee IDs give the same 404 CONTRACT_NOT_FOUND.
      schema:
        type: string
    ContractLimit:
      name: limit
      in: query
      description: Items per page (1-50, default 25).
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 25
...
  schemas:
...
    EmployeeProfile:
      type: object
      required: [id, employeeNumber, givenNames, familyName, businessDate, employment, current]
      properties:
        id:
          type: string
          format: uuid
        employeeNumber:
          type: string
        givenNames:
          type: string
        familyName:
          type: string
        businessDate:
          type: string
          format: date
          description: Today in the organization's time zone; statuses are computed against it.
        employment:
          $ref: "#/components/schemas/Employment"
        current:
          $ref: "#/components/schemas/CurrentAssignments"
...
    ContractPreviewCommand:
      type: object
      additionalProperties: false
      required:
      - templateVersionId
      - startDate
      properties:
        templateVersionId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type:
          - string
          - 'null'
          format: date
    IssueContract:
      type: object
      additionalProperties: false
      required:
      - templateVersionId
      - startDate
      - expectedEmploymentVersion
      - previewDigest
      properties:
        templateVersionId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type:
          - string
          - 'null'
          format: date
        expectedEmploymentVersion:
          type: integer
          minimum: 0
        previewDigest:
          type: string
          pattern: ^[0-9a-f]{64}$
...
    ContractPreview:
      type: object
      required:
      - employmentId
      - employmentVersion
      - templateVersionId
      - contractType
      - locale
      - startDate
      - endDate
      - snapshot
      - integrity
      - warnings
      - previewDigest
      properties:
        employmentId:
          type: string
          format: uuid
        employmentVersion:
          type: integer
          minimum: 0
        templateVersionId:
          type: string
          format: uuid
        contractType:
          $ref: '#/components/schemas/ContractClassification'
        locale:
          $ref: '#/components/schemas/ContractLocale'
        startDate:
          type: string
          format: date
        endDate:
          type:
          - string
          - 'null'
          format: date
        snapshot:
          $ref: '#/components/schemas/ContractSnapshot'
        integrity:
          $ref: '#/components/schemas/ContractIntegrity'
        warnings:
          type: array
          items:
            $ref: '#/components/schemas/ContractWarning'
        previewDigest:
          type: string
          pattern: ^[0-9a-f]{64}$
          description: DIVALHR-CONTRACT-PREVIEW v1 digest (A30-4).
...
    ContractPage:
      type: object
      required:
      - items
      - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ContractSummary'
        nextCursor:
          type:
          - string
          - 'null'
    Contract:
      type: object
      required:
      - id
      - employmentId
      - templateId
      - templateVersionId
      - contractType
      - locale
      - startDate
      - endDate
      - state
      - issuedAt
      - acknowledgedAt
      - voidedAt
      - voidReason
      - version
      - snapshot
      - integrity
      - acknowledgement
      properties:
        id:
          type: string
          format: uuid
        employmentId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        templateVersionId:
          type: string
          format: uuid
        contractType:
          $ref: '#/components/schemas/ContractClassification'
        locale:
          $ref: '#/components/schemas/ContractLocale'
        startDate:
          type: string
          format: date
        endDate:
          type:
          - string
          - 'null'
          format: date
        state:
          $ref: '#/components/schemas/ContractState'
        issuedAt:
          type: string
          format: date-time
        acknowledgedAt:
          type:
          - string
          - 'null'
          format: date-time
        voidedAt:
          type:
          - string
          - 'null'
          format: date-time
        voidReason:
          oneOf:
          - $ref: '#/components/schemas/ContractVoidReason'
          - type: 'null'
        version:
          type: integer
          minimum: 0
        snapshot:
          $ref: '#/components/schemas/ContractSnapshot'
        integrity:
          $ref: '#/components/schemas/ContractIntegrity'
        acknowledgement:
          oneOf:
          - $ref: '#/components/schemas/ContractAcknowledgementEvidence'
          - type: 'null'
    VoidContract:
      type: object
      additionalProperties: false
      required:
      - expectedVersion
      - reasonCode
      properties:
        expectedVersion:
          type: integer
          minimum: 0
        reasonCode:
          $ref: '#/components/schemas/ContractVoidReason'
...